refactor Header add styles in component install styled-components
